Refactoring: Moved classes to individual files
