+++ scanline/ScanlineOutputProcessor.swift
@@ -1,0 +1,142 @@
+//
+//  ScanlineOutputProcessor.swift
+//  scanline
+//
+//  Created by Florian Dreier on 12/30/19.
+//  Copyright © 2019 Scott J. Kleper. All rights reserved.
+//
+
+import Foundation
+import ImageCaptureCore
+import AppKit
+import Quartz
+
+class ScanlineOutputProcessor {
+    let logger: Logger
+    let configuration: ScanConfiguration
+    let urls: [URL]
+    
+    init(urls: [URL], configuration: ScanConfiguration, logger: Logger) {
+        self.urls = urls
+        self.configuration = configuration
+        self.logger = logger
+    }
+    
+    func process() -> Bool {
+        if configuration.config[ScanlineConfigOptionJPEG] != nil {
+            for url in urls {
+                outputAndTag(url: url)
+            }
+        } else {
+            // Combine into a single PDF
+            if let combinedURL = combine(urls: urls) {
+                outputAndTag(url: combinedURL)
+            } else {
+                logger.log("Error while creating PDF")
+                return false
+            }
+        }
+        
+        return true
+    }
+    
+    func combine(urls: [URL]) -> URL? {
+        let document = PDFDocument()
+        
+        for url in urls {
+            if let page = PDFPage(image: NSImage(byReferencing: url)) {
+                document.insert(page, at: document.pageCount)
+            }
+        }
+        
+        let tempFilePath = "\(NSTemporaryDirectory())/scan.pdf"
+        document.write(toFile: tempFilePath)
+        
+        return URL(fileURLWithPath: tempFilePath)
+    }
+
+    func outputAndTag(url: URL) {
+        let gregorian = NSCalendar(calendarIdentifier: .gregorian)!
+        let dateComponents = gregorian.components([.year, .hour, .minute, .second], from: Date())
+        
+        let outputRootDirectory = configuration.config[ScanlineConfigOptionDir] as! String
+        var path = outputRootDirectory
+        
+        // If there's a tag, move the file to the first tag location
+        if configuration.tags.count > 0 {
+            path = "\(path)/\(configuration.tags[0])/\(dateComponents.year!.fld())"
+        }
+        
+        logger.verbose("Output path: \(path)")
+
+        do {
+            try FileManager.default.createDirectory(atPath: path, withIntermediateDirectories: true, attributes: nil)
+        } catch {
+            logger.log("Error while creating directory \(path)")
+            return
+        }
+        let destinationFileExtension = (configuration.config[ScanlineConfigOptionJPEG] != nil ? "jpg" : "pdf")
+        let destinationFileRoot: String = { () -> String in
+            if let fileName = self.configuration.config[ScanlineConfigOptionName] {
+                return "\(path)/\(fileName)"
+            }
+            return "\(path)/scan_\(dateComponents.hour!.f02ld())\(dateComponents.minute!.f02ld())\(dateComponents.second!.f02ld())"
+        }()
+        
+        var destinationFilePath = "\(destinationFileRoot).\(destinationFileExtension)"
+        var i = 0
+        while FileManager.default.fileExists(atPath: destinationFilePath) {
+            destinationFilePath = "\(destinationFileRoot).\(i).\(destinationFileExtension)"
+            i += 1
+        }
+        
+        logger.verbose("About to copy \(url.absoluteString) to \(destinationFilePath)")
+
+        let destinationURL = URL(fileURLWithPath: destinationFilePath)
+        do {
+            try FileManager.default.copyItem(at: url, to: destinationURL)
+        } catch {
+            logger.log("Error while copying file to \(destinationURL.absoluteString)")
+            return
+        }
+
+        // Alias to all other tag locations
+        // todo: this is super repetitive with above...
+        if configuration.tags.count > 1 {
+            for tag in configuration.tags.subarray(with: NSMakeRange(1, configuration.tags.count - 1)) {
+                logger.verbose("Aliasing to tag \(tag)")
+                let aliasDirPath = "\(outputRootDirectory)/\(tag)/\(dateComponents.year!.fld())"
+                do {
+                    try FileManager.default.createDirectory(atPath: aliasDirPath, withIntermediateDirectories: true, attributes: nil)
+                } catch {
+                    logger.log("Error while creating directory \(aliasDirPath)")
+                    return
+                }
+                let aliasFileRoot = { () -> String in
+                    if let name = configuration.config[ScanlineConfigOptionName] {
+                        return "\(aliasDirPath)/\(name)"
+                    }
+                    return "\(aliasDirPath)/scan_\(dateComponents.hour!.f02ld())\(dateComponents.minute!.f02ld())\(dateComponents.second!.f02ld())"
+                }()
+                var aliasFilePath = "\(aliasFileRoot).\(destinationFileExtension)"
+                var i = 0
+                while FileManager.default.fileExists(atPath: aliasFilePath) {
+                    aliasFilePath = "\(aliasFileRoot).\(i).\(destinationFileExtension)"
+                    i += 1
+                }
+                logger.verbose("Aliasing to \(aliasFilePath)")
+                do {
+                    try FileManager.default.createSymbolicLink(atPath: aliasFilePath, withDestinationPath: destinationFilePath)
+                } catch {
+                    logger.log("Error while creating alias at \(aliasFilePath)")
+                    return
+                }
+            }
+        }
+        
+        if configuration.config[ScanlineConfigOptionOpen] != nil {
+            logger.verbose("Opening file at \(destinationFilePath)")
+            NSWorkspace.shared.openFile(destinationFilePath)
+        }
+    }
+}
